Cambios base sincronizacion

+++ SondeoBackend/DTO/Sincronizacion/ProductoDto.cs
@@ -9,6 +9,6 @@ namespace SondeoBackend.DTO.Sincronizacion
         public bool Activado { get; set; }
         public int CategoriaId { get; set; }
         public int MarcaId { get; set; }
-        public int PropiedadesId { get; set; }
+        public long PropiedadesId { get; set; }
     }
 }
