+++ SondeoBackend/Models/DetalleEncuesta.cs
@@ -4,7 +4,7 @@ namespace SondeoBackend.Models
 {
     public class DetalleEncuesta
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public int? StockInicial { get; set; }
         public int? StockFinal { get; set; }
         public float? Compra { get; set; }
@@ -12,7 +12,7 @@ namespace SondeoBackend.Models
         public float? Pvp { get; set; }
         [ForeignKey("EncuestaId")]
         public Encuesta? Encuesta { get; set; }
-        public int? EncuestaId { get; set; }
+        public long? EncuestaId { get; set; }
         [ForeignKey("ProductoId")]
         public Producto? Producto { get; set; }
         public long ProductoId { get; set; }
